photo metadata can now be stored in db

+++ backend/photo/entities/src/main/scala/backend/photo/entities/meta/Coordinates.scala
@@ -1,7 +1,7 @@
 package backend.photo.entities.meta
 
-case class Coordinates(longitude: Int,
-                       latitude: Int) {
+case class Coordinates(longitude: Float,
+                       latitude: Float) {
   require(longitude >= -180 && longitude <= 180, "longitude [-180, 180]")
   require(latitude >= -90 && latitude <= 90, "latitude [-90, 90]")
 }
